--- client/src/hud.rs
@@ -1,13 +1,0 @@
-use wasm_bindgen::prelude::*;
-
-extern crate web_sys;
-
-#[wasm_bindgen]
-pub struct Hud {
-    hud_div: web_sys::Element
-}
-
-#[wasm_bindgen]
-impl Hud {
-
-}
